backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
anthropic==0.42.0
PyMuPDF==1.25.3
playwright==1.49.1
Pillow==11.1.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
